+++ lib/src/separated/separated.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
-import '../utils/text_without_animation.dart';
+
 import '../utils/base_digit_separated.dart';
 import '../utils/enum.dart';
 import '../utils/extensions.dart';
-import '../utils/text_animation.dart';
 import '../utils/separator.dart';
+import '../utils/text_animation.dart';
+import '../utils/text_without_animation.dart';
 
 part 'box_separated.dart';
+
 part 'digit_separated_item.dart';